requirements.txt
numpy
scipy
matplotlib
rich
python-dotenv
schwab-py
